Add files via upload
This simple Java program is designed to assist users with Delhi Metro travel by taking the names of the source and destination stations as input. It displays the fare and the shortest metro route to reach the destination. Additionally, it features a metro map for better navigation. The implementation leverages Graph and Heap data structures, where nodes and edges represent metro stations and the distances between them, respectively. Each node contains information about the station, such as its name, metro corridor, and connecting lines. Edges denote the distance between stations, with the edge cost equal to this distance. By employing algorithms like Dijkstra, breadth-first search, and depth-first search, the program calculates the shortest path and fare based on the total distance. The final output includes the metro route and the total fare.